fix: working on the name conventions
