+++ rudof_cli/src/cli.rs
@@ -1,7 +1,7 @@
 use crate::input_spec::InputSpec;
 use crate::{InputConvertFormat, OutputConvertFormat};
 use clap::{Parser, Subcommand, ValueEnum};
-use shacl_validation::validate::ShaclValidationMode;
+use shacl_validation::shacl_processor::ShaclValidationMode;
 use srdf::{RDFFormat, ReaderMode};
 use std::fmt::Display;
 use std::{fmt::Formatter, path::PathBuf};
