+++ shacl_validation/src/constraints/core/logical/or.rs
@@ -7,9 +7,9 @@ use crate::constraints::constraint_error::ConstraintError;
 use crate::constraints::NativeValidator;
 use crate::constraints::SparqlValidator;
 use crate::constraints::Validator;
-use crate::runner::native::NativeValidatorRunner;
-use crate::runner::sparql::SparqlValidatorRunner;
-use crate::runner::ValidatorRunner;
+use crate::engine::native::NativeEngine;
+use crate::engine::sparql::SparqlEngine;
+use crate::engine::Engine;
 use crate::shape::ShapeValidation;
 use crate::validation_report::result::ValidationResult;
 use crate::validation_report::result::ValidationResults;
@@ -20,7 +20,7 @@ impl<S: SRDFBasic + 'static> Validator<S> for Or<S> {
     fn validate(
         &self,
         store: &S,
-        runner: impl ValidatorRunner<S>,
+        runner: impl Engine<S>,
         value_nodes: &ValueNodes<S>,
     ) -> Result<ValidationResults<S>, ConstraintError> {
         let results = value_nodes
@@ -55,7 +55,7 @@ impl<S: SRDF + 'static> NativeValidator<S> for Or<S> {
         store: &S,
         value_nodes: &ValueNodes<S>,
     ) -> Result<ValidationResults<S>, ConstraintError> {
-        self.validate(store, NativeValidatorRunner, value_nodes)
+        self.validate(store, NativeEngine, value_nodes)
     }
 }
 
@@ -65,6 +65,6 @@ impl<S: QuerySRDF + 'static> SparqlValidator<S> for Or<S> {
         store: &S,
         value_nodes: &ValueNodes<S>,
     ) -> Result<ValidationResults<S>, ConstraintError> {
-        self.validate(store, SparqlValidatorRunner, value_nodes)
+        self.validate(store, SparqlEngine, value_nodes)
     }
 }
