+++ shacl_validation/src/constraints/core/property_pair/equals.rs
@@ -7,9 +7,9 @@ use crate::constraints::constraint_error::ConstraintError;
 use crate::constraints::NativeValidator;
 use crate::constraints::SparqlValidator;
 use crate::constraints::Validator;
-use crate::runner::native::NativeValidatorRunner;
-use crate::runner::sparql::SparqlValidatorRunner;
-use crate::runner::ValidatorRunner;
+use crate::engine::native::NativeEngine;
+use crate::engine::sparql::SparqlEngine;
+use crate::engine::Engine;
 use crate::validation_report::result::ValidationResults;
 use crate::ValueNodes;
 
@@ -17,7 +17,7 @@ impl<S: SRDFBasic + 'static> Validator<S> for Equals<S> {
     fn validate(
         &self,
         _store: &S,
-        _runner: impl ValidatorRunner<S>,
+        _runner: impl Engine<S>,
         _value_nodes: &ValueNodes<S>,
     ) -> Result<ValidationResults<S>, ConstraintError> {
         Err(ConstraintError::NotImplemented)
@@ -30,7 +30,7 @@ impl<S: SRDF + 'static> NativeValidator<S> for Equals<S> {
         store: &S,
         value_nodes: &ValueNodes<S>,
     ) -> Result<ValidationResults<S>, ConstraintError> {
-        self.validate(store, NativeValidatorRunner, value_nodes)
+        self.validate(store, NativeEngine, value_nodes)
     }
 }
 
@@ -40,6 +40,6 @@ impl<S: QuerySRDF + 'static> SparqlValidator<S> for Equals<S> {
         store: &S,
         value_nodes: &ValueNodes<S>,
     ) -> Result<ValidationResults<S>, ConstraintError> {
-        self.validate(store, SparqlValidatorRunner, value_nodes)
+        self.validate(store, SparqlEngine, value_nodes)
     }
 }
