+++ shacl_validation/src/constraints/core/string_based/unique_lang.rs
@@ -10,9 +10,9 @@ use crate::constraints::constraint_error::ConstraintError;
 use crate::constraints::NativeValidator;
 use crate::constraints::SparqlValidator;
 use crate::constraints::Validator;
-use crate::runner::native::NativeValidatorRunner;
-use crate::runner::sparql::SparqlValidatorRunner;
-use crate::runner::ValidatorRunner;
+use crate::engine::native::NativeEngine;
+use crate::engine::sparql::SparqlEngine;
+use crate::engine::Engine;
 use crate::validation_report::result::ValidationResult;
 use crate::validation_report::result::ValidationResults;
 use crate::ValueNodes;
@@ -21,7 +21,7 @@ impl<S: SRDFBasic + 'static> Validator<S> for UniqueLang {
     fn validate(
         &self,
         _: &S,
-        _: impl ValidatorRunner<S>,
+        _: impl Engine<S>,
         value_nodes: &ValueNodes<S>,
     ) -> Result<ValidationResults<S>, ConstraintError> {
         if !self.unique_lang() {
@@ -63,7 +63,7 @@ impl<S: SRDF + 'static> NativeValidator<S> for UniqueLang {
         store: &S,
         value_nodes: &ValueNodes<S>,
     ) -> Result<ValidationResults<S>, ConstraintError> {
-        self.validate(store, NativeValidatorRunner, value_nodes)
+        self.validate(store, NativeEngine, value_nodes)
     }
 }
 
@@ -73,6 +73,6 @@ impl<S: QuerySRDF + 'static> SparqlValidator<S> for UniqueLang {
         store: &S,
         value_nodes: &ValueNodes<S>,
     ) -> Result<ValidationResults<S>, ConstraintError> {
-        self.validate(store, SparqlValidatorRunner, value_nodes)
+        self.validate(store, SparqlEngine, value_nodes)
     }
 }
