+++ shacl_validation/src/engine/sparql.rs
@@ -1,25 +1,22 @@
+use indoc::formatdoc;
 use shacl_ast::compiled::component::CompiledComponent;
 use shacl_ast::compiled::property_shape::CompiledPropertyShape;
 use shacl_ast::compiled::shape::CompiledShape;
+use srdf::QuerySRDF;
 use srdf::SHACLPath;
-use srdf::RDFS_CLASS;
-use srdf::RDFS_SUBCLASS_OF;
-use srdf::RDF_TYPE;
-use srdf::SRDF;
 
-use crate::constraints::NativeDeref;
-use crate::helper::srdf::get_objects_for;
-use crate::helper::srdf::get_subjects_for;
+use crate::constraints::SparqlDeref;
+use crate::helper::sparql::select;
 use crate::validate_error::ValidateError;
 use crate::validation_report::result::ValidationResults;
 use crate::Targets;
 use crate::ValueNodes;
 
-use super::ValidatorRunner;
+use super::Engine;
 
-pub struct NativeValidatorRunner;
+pub struct SparqlEngine;
 
-impl<S: SRDF + 'static> ValidatorRunner<S> for NativeValidatorRunner {
+impl<S: QuerySRDF + 'static> Engine<S> for SparqlEngine {
     fn evaluate(
         &self,
         store: &S,
@@ -27,17 +24,26 @@ impl<S: SRDF + 'static> ValidatorRunner<S> for NativeValidatorRunner {
         value_nodes: &ValueNodes<S>,
     ) -> Result<ValidationResults<S>, ValidateError> {
         let validator = component.deref();
-        Ok(validator.validate_native(store, value_nodes)?)
+        Ok(validator.validate_sparql(store, value_nodes)?)
     }
 
     /// If s is a shape in a shapes graph SG and s has value t for sh:targetNode
     /// in SG then { t } is a target from any data graph for s in SG.
-    fn target_node(&self, _: &S, node: &S::Term) -> Result<Targets<S>, ValidateError> {
+    fn target_node(&self, store: &S, node: &S::Term) -> Result<Targets<S>, ValidateError> {
         if S::term_is_bnode(node) {
-            Err(ValidateError::TargetNodeBlankNode)
-        } else {
-            Ok(Targets::new(std::iter::once(node.clone())))
+            return Err(ValidateError::TargetNodeBlankNode);
         }
+
+        let query = formatdoc! {"
+            SELECT DISTINCT ?this
+            WHERE {{
+                BIND ({} AS ?this)
+            }}
+        ", node};
+
+        select(store, query, "this")?;
+
+        Err(ValidateError::NotImplemented)
     }
 
     fn target_class(&self, store: &S, class: &S::Term) -> Result<Targets<S>, ValidateError> {
@@ -45,14 +51,19 @@ impl<S: SRDF + 'static> ValidatorRunner<S> for NativeValidatorRunner {
             return Err(ValidateError::TargetClassNotIri);
         }
 
-        let subjects = match store.subjects_with_predicate_object(&S::iri_s2iri(&RDF_TYPE), class) {
-            Ok(subjects) => subjects,
-            Err(_) => return Err(ValidateError::SRDF),
-        };
+        let query = formatdoc! {"
+            PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
+            PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>
 
-        let targets = subjects.iter().map(|subject| S::subject_as_term(subject));
+            SELECT DISTINCT ?this
+            WHERE {{
+                ?this rdf:type/rdfs:subClassOf* {} .
+            }}
+        ", class};
 
-        Ok(Targets::new(targets))
+        select(store, query, "this")?;
+
+        Err(ValidateError::NotImplemented)
     }
 
     fn target_subject_of(
@@ -60,73 +71,47 @@ impl<S: SRDF + 'static> ValidatorRunner<S> for NativeValidatorRunner {
         store: &S,
         predicate: &S::IRI,
     ) -> Result<Targets<S>, ValidateError> {
-        let triples = match store.triples_with_predicate(predicate) {
-            Ok(triples) => triples,
-            Err(_) => return Err(ValidateError::SRDF),
-        };
+        let query = formatdoc! {"
+            SELECT DISTINCT ?this
+            WHERE {{
+                ?this {} ?any .
+            }}
+        ", predicate};
 
-        let targets = triples
-            .iter()
-            .map(|triple| S::subject_as_term(&triple.subj()));
+        select(store, query, "this")?;
 
-        Ok(Targets::new(targets))
+        Err(ValidateError::NotImplemented)
     }
 
     fn target_object_of(&self, store: &S, predicate: &S::IRI) -> Result<Targets<S>, ValidateError> {
-        let triples = match store.triples_with_predicate(predicate) {
-            Ok(triples) => triples,
-            Err(_) => return Err(ValidateError::SRDF),
-        };
+        let query = formatdoc! {"
+            SELECT DISTINCT ?this
+            WHERE {{
+                ?any {} ?this .
+            }}
+        ", predicate};
 
-        let targets = triples.into_iter().map(|triple| triple.obj());
+        select(store, query, "this")?;
 
-        Ok(Targets::new(targets))
+        Err(ValidateError::NotImplemented)
     }
 
     fn implicit_target_class(
         &self,
-        store: &S,
-        shape: &CompiledShape<S>,
+        _store: &S,
+        _shape: &CompiledShape<S>,
     ) -> Result<Targets<S>, ValidateError> {
-        let ctypes = get_objects_for(store, shape.id(), &S::iri_s2iri(&RDF_TYPE))?;
-
-        let mut subclasses = get_subjects_for(
-            store,
-            &S::iri_s2iri(&RDFS_SUBCLASS_OF),
-            &S::iri_s2term(&RDFS_CLASS),
-        )?;
-
-        subclasses.insert(S::iri_s2term(&RDFS_CLASS));
-
-        if ctypes.iter().any(|t| subclasses.contains(t)) {
-            let actual_class_nodes = get_subjects_for(store, &S::iri_s2iri(&RDF_TYPE), shape.id())?;
-
-            let subclass_targets =
-                get_subjects_for(store, &S::iri_s2iri(&RDFS_SUBCLASS_OF), shape.id())?
-                    .into_iter()
-                    .flat_map(move |subclass| {
-                        get_subjects_for(store, &S::iri_s2iri(&RDF_TYPE), &subclass)
-                            .into_iter()
-                            .flatten()
-                    });
-
-            let targets = actual_class_nodes.into_iter().chain(subclass_targets);
-            Ok(Targets::new(targets))
-        } else {
-            Ok(Targets::default())
-        }
+        Err(ValidateError::NotImplemented)
     }
 
     fn predicate(
         &self,
-        store: &S,
-        _: &CompiledPropertyShape<S>,
-        predicate: &S::IRI,
-        focus_node: &S::Term,
+        _store: &S,
+        _shape: &CompiledPropertyShape<S>,
+        _predicate: &S::IRI,
+        _focus_node: &S::Term,
     ) -> Result<Targets<S>, ValidateError> {
-        Ok(Targets::new(
-            get_objects_for(store, focus_node, predicate)?.into_iter(),
-        ))
+        Err(ValidateError::NotImplemented)
     }
 
     fn alternative(
