+++ shacl_validation/src/lib.rs
@@ -4,13 +4,13 @@ use std::collections::HashSet;
 use srdf::SRDFBasic;
 
 pub(crate) mod constraints;
+pub(crate) mod engine;
 pub mod helper;
-pub(crate) mod runner;
 pub mod shacl_config;
+pub mod shacl_processor;
 pub mod shacl_validation_vocab;
 pub(crate) mod shape;
 pub mod store;
-pub mod validate;
 pub mod validate_error;
 pub mod validation_report;
