+++ shacl_validation/src/shacl_processor.rs
@@ -7,9 +7,9 @@ use srdf::SRDFBasic;
 use srdf::SRDFGraph;
 use srdf::SRDFSparql;
 
-use crate::runner::native::NativeValidatorRunner;
-use crate::runner::sparql::SparqlValidatorRunner;
-use crate::runner::ValidatorRunner;
+use crate::engine::native::NativeEngine;
+use crate::engine::sparql::SparqlEngine;
+use crate::engine::Engine;
 use crate::shape::ShapeValidation;
 use crate::store::graph::Graph;
 use crate::store::sparql::Endpoint;
@@ -28,13 +28,13 @@ pub enum ShaclValidationMode {
     Sparql,
 }
 
-/// The Validation trait is the one in charge of applying the SHACL Validation
+/// The ShaclProcessor trait is the one in charge of applying the SHACL Validation
 /// algorithm. For this, first, the validation report is initiliazed to empty,
 /// and, for each shape in the schema, the target nodes are selected, and then,
 /// each validator for each constraint is applied.
-pub trait Validation<S: SRDFBasic> {
+pub trait ShaclProcessor<S: SRDFBasic> {
     fn store(&self) -> &S;
-    fn runner(&self) -> &dyn ValidatorRunner<S>;
+    fn runner(&self) -> &dyn Engine<S>;
 
     fn validate(&self, schema: CompiledSchema<S>) -> Result<ValidationReport<S>, ValidateError> {
         // we initialize the validation report to empty
@@ -73,14 +73,14 @@ impl GraphValidation {
     }
 }
 
-impl Validation<SRDFGraph> for GraphValidation {
+impl ShaclProcessor<SRDFGraph> for GraphValidation {
     fn store(&self) -> &SRDFGraph {
         self.store.store()
     }
 
-    fn runner(&self) -> &dyn ValidatorRunner<SRDFGraph> {
+    fn runner(&self) -> &dyn Engine<SRDFGraph> {
         match self.mode {
-            ShaclValidationMode::Native => &NativeValidatorRunner,
+            ShaclValidationMode::Native => &NativeEngine,
             ShaclValidationMode::Sparql => todo!(),
         }
     }
@@ -100,15 +100,15 @@ impl EndpointValidation {
     }
 }
 
-impl Validation<SRDFSparql> for EndpointValidation {
+impl ShaclProcessor<SRDFSparql> for EndpointValidation {
     fn store(&self) -> &SRDFSparql {
         self.store.store()
     }
 
-    fn runner(&self) -> &dyn ValidatorRunner<SRDFSparql> {
+    fn runner(&self) -> &dyn Engine<SRDFSparql> {
         match self.mode {
-            ShaclValidationMode::Native => &NativeValidatorRunner,
-            ShaclValidationMode::Sparql => &SparqlValidatorRunner,
+            ShaclValidationMode::Native => &NativeEngine,
+            ShaclValidationMode::Sparql => &SparqlEngine,
         }
     }
 }
